+++ src/main/java/ch/alpine/tensor/pdf/d/GeometricDistribution.java
@@ -76,7 +76,7 @@ public class GeometricDistribution extends AbstractDiscreteDistribution implemen
   }
 
   @Override // from AbstractDiscreteDistribution
-  protected Scalar protected_p_equals(int x) {
+  protected Scalar protected_p_equals(BigInteger x) {
     // PDF[GeometricDistribution[p], x] == (1 - p) ^ x * p
     return p.multiply(Power.of(_1_p, x));
   }
